move JWTSource to separate file
